Added racial skill bonuses and randomization data to DB; added names of deities and characters in plaintext; improved search

+++ character.php
@@ -73,7 +73,7 @@ if (!$_SESSION["auth"]) {
 
 <!--		<a href="item.php?char=--><?php //echo $charId; ?><!--">Add Item</a> |-->
 		<a href="character_form.php?mode=edit&char=<?= $charId; ?>">Edit Character</a> |
-		<a href="delete_character.php?char=<?= $charId; ?>">Delete Character</a>
+		<a href="delete_character.php?char=<?= $charId; ?>" onclick="return confirm('Are you sure? you want to delete this character?')">Delete Character</a>
 		<?php
 		// basic modifier calculations
 		$str_mod = (floor($row["str_attr"]/2)-5);
